+++ demisto_sdk/commands/unify/unifier.py
@@ -2,6 +2,7 @@ import base64
 import copy
 import glob
 import io
+import json
 import os
 import re
 from typing import Tuple
@@ -19,6 +20,11 @@ from demisto_sdk.commands.common.tools import (LOG_COLORS, find_type, get_yaml,
                                                server_version_compare)
 from ruamel.yaml import YAML
 from ruamel.yaml.scalarstring import FoldedScalarString
+
+PACK_METADATA_PATH = 'pack_metadata.json'
+PARTNER_CONTRIB_DISPLAY_NAME = ' (Partner contribution)'
+PARTNER_CONTRIB_DETAILED_DESC = '### This is a Partner Contributed integration' \
+                                '\nFor all questions and enhancement requests please contact the partner directly:'
 
 
 class Unifier:
@@ -156,6 +162,11 @@ class Unifier:
         if not self.is_script_package:
             yml_unified, image_path = self.insert_image_to_yml(self.yml_data, yml_unified)
             yml_unified, desc_path = self.insert_description_to_yml(self.yml_data, yml_unified)
+            is_parnter_contrib, metadata_data = self.is_partner_contributed()
+            if is_parnter_contrib:
+                partner_email = metadata_data.get('email', '')
+                partner_url = metadata_data.get('url', '')
+                yml_unified = self.add_partner_support(yml_unified, partner_email, partner_url)
 
         output_map = self.write_yaml_with_docker(yml_unified, self.yml_data, script_obj)
         unifier_outputs = list(output_map.keys()), self.yml_path, script_path, image_path, desc_path
@@ -164,7 +175,7 @@ class Unifier:
         return unifier_outputs[0]
 
     def insert_image_to_yml(self, yml_data, yml_unified):
-        image_data, found_img_path = self.get_data("*png")
+        image_data, found_img_path = self.get_data(self.package_path, "*png")
         image_data = self.image_prefix + base64.b64encode(image_data).decode('utf-8')
 
         if yml_data.get('image') and self.use_force is False:
@@ -176,7 +187,7 @@ class Unifier:
         return yml_unified, found_img_path
 
     def insert_description_to_yml(self, yml_data, yml_unified):
-        desc_data, found_desc_path = self.get_data('*_description.md')
+        desc_data, found_desc_path = self.get_data(self.package_path, '*_description.md')
 
         if yml_data.get('detaileddescription') and self.use_force is False:
             raise ValueError('Please move the detailed description from the yml to a description file (.md)'
@@ -186,8 +197,8 @@ class Unifier:
 
         return yml_unified, found_desc_path
 
-    def get_data(self, extension):
-        data_path = glob.glob(os.path.join(self.package_path, extension))
+    def get_data(self, path, extension):
+        data_path = glob.glob(os.path.join(path, extension))
         data = None
         found_data_path = None
         if not self.is_script_package and data_path:
@@ -338,3 +349,44 @@ class Unifier:
         script_code = script_code.replace(". $PSScriptRoot\\demistomock.ps1", "")
         script_code = script_code.replace(". $PSScriptRoot\\CommonServerPowerShell.ps1", "")
         return script_code
+
+    def get_pack_path(self):
+        return self.package_path.split('Integrations')[0]
+
+    def is_partner_contributed(self):
+        """Checks if the integration is a partner contribution.
+
+        Returns:
+            (bool, dict). True if the integration is a partner contribution and the file data, False and None otherwise.
+        """
+        pack_path = self.get_pack_path()
+
+        pack_metadata_data, pack_metadata_path = self.get_data(pack_path, PACK_METADATA_PATH)
+        if pack_metadata_data:
+            json_pack_metadata = json.loads(pack_metadata_data)
+            support_field = json_pack_metadata.get('support')
+            if support_field == 'partner':
+                return True, json_pack_metadata
+        return False, None
+
+    def add_partner_support(self, unified_yml, partner_email, partner_url):
+        """Add partner support to the unified file - text in the display name and detailed description.
+
+        Args:
+            unified_yml: The unified yaml file.
+            partner_email (str): The partner email.
+            partner_url (str): The partner url.
+
+        Returns:
+            The unified yaml file.
+        """
+        unified_yml['display'] += PARTNER_CONTRIB_DISPLAY_NAME
+        existing_detailed_description = unified_yml['detaileddescription']
+        partner_description = PARTNER_CONTRIB_DETAILED_DESC
+        if partner_email:
+            partner_description += f'\n**Email** - {partner_email}'
+        if partner_url:
+            partner_description += f'\n**URL** - {partner_url}'
+        unified_yml['detaileddescription'] = partner_description + '\n***\n' + existing_detailed_description
+
+        return unified_yml